+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Added: Element#getExtendedData - without table name, to get all extended data rows for that element 
 * Added: Fetch hints to get old metadata
 * Fixed: Elasticsearch query strings with field names - handling of values other than constants
+* Changed: Speed up element delete when no extended data rows are present
 
 # v4.6.0
 * Added: History for vertices, edges, and properties